Deploy/requirements.txt
altair==4.2.0
anyio==3.6.1
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorlover==0.3.0
commonmark==0.9.1
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
investpy==1.0.8
ipykernel==6.15.2
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mplfinance==0.12.9b1
multitasking==0.0.11
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-bcb==0.1.8
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
requests==2.28.1
rfc3986==1.5.0
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.5.0
streamlit==1.12.2
tenacity==8.0.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
widgetsnbextension==4.0.3
yfinance==0.1.74
zipp==3.8.1